refactor: moving useEffect/fn's, changing icon and file names
