+++ src/app/layout.tsx
@@ -11,6 +11,7 @@ const inter = Inter({ subsets: ["latin"] });
 
 export const metadata: Metadata = {
   title: "Create Next App",
+  icons: 'favico.png',
   description: "Generated by create next app",
 };
 
@@ -20,7 +21,9 @@ export default function RootLayout({
   children: React.ReactNode;
 }) {
   return (
+    
     <html lang="en">
+
       <StoreProvider>
         <body className={inter.className}>
           <AppRouterCacheProvider>   <ThemeProvider theme={theme}>{children}</ThemeProvider></AppRouterCacheProvider>
