+++ src/app/page.tsx
@@ -1,6 +1,6 @@
 import { sql } from "@vercel/postgres";
 import { unstable_noStore as noStore } from "next/cache";
-import FindConnection from "../components/(user)/home/connection";
+import SearchConnection from "../components/(user)/home/searchConnection";
 import Modal from "../components/(global)/modal";
 export default async function Home() {
   noStore();
@@ -10,7 +10,7 @@ export default async function Home() {
     <main className="flex text-black  min-h-screen flex-col items-center py-3 px-3">
       <Modal></Modal>
       <h1 className="text-2xl">Find connection</h1>
-      <FindConnection airports={airports as Airport[]}></FindConnection>
+      <SearchConnection airports={airports as Airport[]}></SearchConnection>
    
     </main>
   );
